# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mavericks Inventory API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StockItem {
  id: number;
  stockCode: string;
  stockName: string;
  category: string;
  unitOfMeasure: string;
  openingQuantity: number;
  availableQuantity: number;
  /** @nullable */
  minStockLevel?: number | null;
  /** @nullable */
  location: string | null;
  status: string;
  /** @nullable */
  healthScore?: number | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface StockItemInput {
  stockCode: string;
  stockName: string;
  category: string;
  unitOfMeasure: string;
  openingQuantity: number;
  minStockLevel?: number;
  location?: string;
  status?: string;
}

export interface StockItemUpdate {
  stockName?: string;
  category?: string;
  unitOfMeasure?: string;
  location?: string;
  minStockLevel?: number;
  status?: string;
}

export interface LedgerEntry {
  id: number;
  stockId: number;
  type: string;
  quantity: number;
  description: string;
  createdAt: string;
  /** @nullable */
  createdBy?: string | null;
}

export interface Distribution {
  id: number;
  stockId: number;
  stockName: string;
  /** @nullable */
  stockCode?: string | null;
  quantityDistributed: number;
  distributionDate: string;
  recipient: string;
  /** @nullable */
  purpose: string | null;
  status: string;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  createdAt: string;
}

export interface DistributionInput {
  stockId: number;
  quantityDistributed: number;
  distributionDate: string;
  recipient: string;
  purpose?: string;
}

export interface DistributionUpdate {
  quantityDistributed?: number;
  distributionDate?: string;
  recipient?: string;
  purpose?: string;
}

export interface AiRecommendation {
  recommendation: string;
  riskScore: number;
  riskLevel: string;
  reasoning: string;
  /** @nullable */
  confidence?: number | null;
}

export interface ApprovalRequest {
  id: number;
  distributionId: number;
  distribution: Distribution;
  status: string;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  approvedBy?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  aiRecommendation?: AiRecommendation;
  createdAt: string;
}

export interface ApprovalAction {
  remarks: string;
}

export interface DashboardSummary {
  totalStockItems: number;
  totalAvailableUnits: number;
  pendingApprovals: number;
  anomalyCount: number;
  healthyCount: number;
  warningCount: number;
  criticalCount: number;
  totalDistributions: number;
  /** @nullable */
  automationPercentage?: number | null;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  actor?: string | null;
  /** @nullable */
  stockCode?: string | null;
}

export interface StockHealthScore {
  stockId: number;
  stockCode: string;
  stockName: string;
  healthScore: number;
  healthStatus: string;
  availableQuantity: number;
  category: string;
  /** @nullable */
  location?: string | null;
}

export interface AnomalyAlert {
  id: number;
  stockId: number;
  stockCode: string;
  stockName: string;
  type: string;
  severity: string;
  description: string;
  /** @nullable */
  explanation?: string | null;
  detectedAt: string;
  dismissed: boolean;
  status: string;
  /** @nullable */
  acknowledgedBy?: string | null;
  /** @nullable */
  acknowledgedAt?: string | null;
  /** @nullable */
  resolvedBy?: string | null;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  resolutionNotes?: string | null;
}

export interface InventoryInsight {
  summary: string;
  generatedAt: string;
  healthPercentage: number;
  keyObservations?: string[];
  recommendedActions?: string[];
}

export interface InsightQuery {
  query: string;
}

export type InsightQueryResponseDataItem = { [key: string]: unknown };

export interface InsightQueryResponse {
  answer: string;
  queryType: string;
  generatedAt: string;
  data?: InsightQueryResponseDataItem[];
}

export type ListStocksParams = {
category?: string;
location?: string;
status?: string;
search?: string;
};

export type ListDistributionsParams = {
status?: string;
stockId?: number;
};

export type ListApprovalsParams = {
status?: string;
};

export type ListAnomaliesParams = {
all?: string;
status?: string;
};

export type AcknowledgeAnomalyBody = {
  acknowledgedBy: string;
};

export type ResolveAnomalyBody = {
  notes: string;
  resolvedBy?: string;
};

export type ListActivityParams = {
type?: string;
actor?: string;
limit?: number;
};