# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mavericks Inventory API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List pending approval requests
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Approval requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalRequest"
  /approvals/{id}:
    get:
      operationId: getApproval
      tags: [approvals]
      summary: Get an approval request with AI recommendation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approval request with AI data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
        "404":
          description: Not found
  /approvals/{id}/approve:
    post:
      operationId: approveRequest
      tags: [approvals]
      summary: Approve a distribution request
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
  /approvals/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [approvals]
      summary: Reject a distribution request
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"

  # ── Dashboard ─────────────────────────────────────────────────
...
  /anomalies:
    get:
      operationId: listAnomalies
      tags: [anomalies]
      summary: List anomaly alerts
      parameters:
        - in: query
          name: all
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Anomaly alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnomalyAlert"
  /anomalies/{id}/dismiss:
    post:
      operationId: dismissAnomaly
      tags: [anomalies]
      summary: Dismiss an anomaly alert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnomalyAlert"
  /anomalies/{id}/acknowledge:
    post:
      operationId: acknowledgeAnomaly
      tags: [anomalies]
      summary: Acknowledge an anomaly alert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [acknowledgedBy]
              properties:
                acknowledgedBy: { type: string }
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnomalyAlert"
  /anomalies/{id}/resolve:
    post:
      operationId: resolveAnomaly
      tags: [anomalies]
      summary: Resolve an anomaly alert with notes
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notes]
              properties:
                notes: { type: string }
                resolvedBy: { type: string }
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnomalyAlert"

  # ── Activity / Audit Log ─────────────────────────────────────
...
components:
  schemas:
...
    Distribution:
      type: object
      required: [id, stockId, stockName, quantityDistributed, distributionDate, recipient, purpose, status, createdAt]
      properties:
        id:                   { type: integer }
        stockId:              { type: integer }
        stockName:            { type: string }
        stockCode:            { type: ["string", "null"] }
        quantityDistributed:  { type: number }
        distributionDate:     { type: string }
        recipient:            { type: string }
        purpose:              { type: ["string", "null"] }
        status:               { type: string }
        remarks:              { type: ["string", "null"] }
        submittedAt:          { type: ["string", "null"] }
        approvedAt:           { type: ["string", "null"] }
        createdAt:            { type: string }
...
    ApprovalRequest:
      type: object
      required: [id, distributionId, distribution, status, createdAt]
      properties:
        id:             { type: integer }
        distributionId: { type: integer }
        distribution:   { $ref: "#/components/schemas/Distribution" }
        status:         { type: string }
        remarks:        { type: ["string", "null"] }
        approvedBy:     { type: ["string", "null"] }
        approvedAt:     { type: ["string", "null"] }
        aiRecommendation: { $ref: "#/components/schemas/AiRecommendation" }
        createdAt:      { type: string }
    AiRecommendation:
      type: object
      required: [recommendation, riskScore, riskLevel, reasoning]
      properties:
        recommendation: { type: string }
        riskScore:      { type: number }
        riskLevel:      { type: string }
        reasoning:      { type: string }
        confidence:     { type: ["number", "null"] }
    ApprovalAction:
      type: object
      required: [remarks]
      properties:
        remarks: { type: string }
...
    AnomalyAlert:
      type: object
      required: [id, stockId, stockCode, stockName, type, severity, description, detectedAt, dismissed, status]
      properties:
        id:               { type: integer }
        stockId:          { type: integer }
        stockCode:        { type: string }
        stockName:        { type: string }
        type:             { type: string }
        severity:         { type: string }
        description:      { type: string }
        explanation:      { type: ["string", "null"] }
        detectedAt:       { type: string }
        dismissed:        { type: boolean }
        status:           { type: string }
        acknowledgedBy:   { type: ["string", "null"] }
        acknowledgedAt:   { type: ["string", "null"] }
        resolvedBy:       { type: ["string", "null"] }
        resolvedAt:       { type: ["string", "null"] }
        resolutionNotes:  { type: ["string", "null"] }
...