# lib/api-zod/src/generated/types/anomalyAlert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mavericks Inventory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnomalyAlert {
  id: number;
  stockId: number;
  stockCode: string;
  stockName: string;
  type: string;
  severity: string;
  description: string;
  /** @nullable */
  explanation?: string | null;
  detectedAt: string;
  dismissed: boolean;
  status: string;
  /** @nullable */
  acknowledgedBy?: string | null;
  /** @nullable */
  acknowledgedAt?: string | null;
  /** @nullable */
  resolvedBy?: string | null;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  resolutionNotes?: string | null;
}