# lib/api-zod/src/generated/types/distribution.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mavericks Inventory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Distribution {
  id: number;
  stockId: number;
  stockName: string;
  /** @nullable */
  stockCode?: string | null;
  quantityDistributed: number;
  distributionDate: string;
  recipient: string;
  /** @nullable */
  purpose: string | null;
  status: string;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  createdAt: string;
}